Interface:  - Population code added  - Adapted to work with index between 0,1 or with integer index like population code All index are implemented now for one layer and for multiple layers. Only specific neuron windows remain

+++ nefesi/layer_data.py
@@ -116,7 +116,7 @@ class LayerData(object):
                     print(self.layer_id+": "+str(i)+"/"+str(len(self.neurons_data)))
                 sel_idx[i] = self.neurons_data[i].class_selectivity_idx(labels, thr_class_idx)
         elif index_name.lower() == 'population code':
-            sel_idx = np.zeros(len(self.neurons_data), dtype=np.float)
+            sel_idx = np.zeros(len(self.neurons_data), dtype=np.int)
             for i in range(len(self.neurons_data)):
                 if verbose:
                     print(self.layer_id+": "+str(i)+"/"+str(len(self.neurons_data)))
